+++ classes/mage.lua
@@ -33,6 +33,7 @@ local HotStreakHandler = {}
 -- Initialize constants
 HotStreakHandler.init = function(self, talentName)
     local fire_blast = { 2136, 2137, 2138, 8412, 8413, 10197, 10199, 27078, 27079, 42872, 42873 }
+    local fire_blast_sod = { 400618, 400619, 400616, 400620, 400621, 400622, 400623 } -- Improved by Overheat rune
     local fireball = { 133, 143, 145, 3140, 8400, 8401, 8402, 10148, 10149, 10150, 10151, 25306, 27070, 38692, 42832, 42833 }
     local frostfire_bolt = { 44614, 47610 }
     local frostfire_bolt_sod = { 401502 }
@@ -51,6 +52,7 @@ HotStreakHandler.init = function(self, talentName)
         end
     end
     addSpellPack(fire_blast);
+    addSpellPack(fire_blast_sod);
     addSpellPack(fireball);
     addSpellPack(frostfire_bolt);
     addSpellPack(frostfire_bolt_sod);