+++ changelog.md
@@ -3,6 +3,7 @@
 #### v2.3.1 (2024-07-xx)
 
 - Mage's Hot Streak triggers from Balefire Bolt (Season of Discovery)
+- Mage's Hot Streak triggers from Overheat's Fire Blast (SoD)
 
 Bug Fixes
 - Mage's Hot Streak was not detected correctly since last patch on Classic Era