+++ changelog.md
@@ -4,6 +4,7 @@
 
 New rune for Season of Discovery:
 - New SAO: Warlock's Decimation
+- New GAB: Warlock's Drain Soul, with Soul Siphon rune
 - New GAB: Warlock's Soul Fire, during Decimation
 
 Updated runes for Season of Discovery